backend/requirements.txt
# Design Ref: §11.2 — Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# AI / LangGraph
langgraph==0.2.60
langchain-openai==0.2.14
langchain-core==0.3.28
openai==1.58.0

# Supabase
supabase==2.11.0

# TTS
elevenlabs==1.15.0

# FFmpeg
ffmpeg-python==0.2.0

# Auth
python-jose[cryptography]==3.3.0

# Database (Checkpointer)
langgraph-checkpoint-postgres==2.0.10
asyncpg==0.30.0
psycopg[binary]==3.2.4

# Monitoring
sentry-sdk[fastapi]==2.19.0

# Utilities
httpx==0.28.0
python-multipart==0.0.18
redis==5.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
